backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
zeep>=4.3.0
lxml>=5.3.0
cryptography>=43.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
Pillow>=10.4.0
